Latest version...

+++ timing_test/tic_cmd_measurement_test.h
@@ -6,6 +6,8 @@
 #ifndef TIC_CMD_MEASUREMENT_TEST_H
 #define TIC_CMD_MEASUREMENT_TEST_H
 
-void tic_cmd_measurement_test(unsigned repeats, unsigned min_cycle_count, unsigned max_cycle_count);
+#include <stdint.h>
+
+void tic_cmd_measurement_test(unsigned repeats, uint8_t min_cycle_count, uint8_t max_cycle_count);
 
 #endif
